Build Cube compooent
This commit adds basic functionallity to the Cube component.
It will receive an array of 6 letter' (noting that
Qu is a possible 'letter') from its parent Board component,
and will select one of them to display on the cube. This selection
is made using the custom `useSingleton` hook; credit to
https://dev.to/bytebodger/constructors-in-functional-components-with-hooks-280m