Build CSS to display Cubes
This commit also modifies the Game component to wrap the imported
Board component in a div with class Game. For the CSS changes,
this commit also introduces an intermediate Cube-container
class for the li items containing the Cube react component,
so that each cube lives in its correct 1/16th of the board
and can still have a reasonable border (5% margin at this time).

The CSS is also updated to center the letter text; see
https://www.w3schools.com/css/tryit.asp?filename=trycss_align_transform