Build WordList display component
This commit adds a new WordListContext for allowing communication
between the ButtonContainer and WordList component, so that the currentWord
can be added to the wordList and displayed.

It remains to implement restrictions on words that can be submitted, as
well as preventing the submission of a blank word (the empty string).