Add coordinates to Cube components
The Board component previously fetched the cube letters and randomized
the cube order, then producing a ul of Cube components. The
Board component now additionally encodes the coordinates (row, column)
of each cube. As such, the Cube component also takes the
row and column props as inputs.

This is part of an ongoing effort to make sure that only valid
cubes (by adjacency) can be seleced to make potential words.