Update Interface to accept new Cube format
The previous event handlers associated with submitting, clearing,
and displaying the current word are updated in this commit to
accept the new format of a Cube as having relevant row and column
information, so that the currentWord context now houses
the currentWord as an array of letterData objects {letter, row, column}.
As such, this commut introduces the parseCurrentWord function
to pull the letter from the letterData object and display, submit, and
clear as needed.

It remains to create and implement a distance-checking function
between cubes on click.