Prevent re-using boggle cubes
A previous commit (a9bcddb) added the index prop to the Cube
component; this commit adds a check in the Cube
component to ensure that a Cube which has been added to the
currentWord context may not be used a second time. In the
event that a Cube is clicked more than once (that is, clicked
again after it being added to the currentWord), a
message is logged "cube already used".