Move WordLlistContext to Game
This commit simplifies the context structure by moving the
WordList state and corresponding context into the Game
component. To accommodate this change, the WordList is now
a stateless component, accepting a single wordList prop.
From the perspective of the ButtonContainer component, the
only change is the source of the import for the
WordListContext.

While not directly related to this issue, this commit works
in the spirit of #1.