Improve selectedCubes styling logic
This commit removes the Cube-level logic which previously reset the
currentCube in the case that a cube was clicked twice in a row.
Instead, the Game now automatically resets the current cube
(to `{index: -1}`) after each useEffect run.

Now, the selectedCubes array stores the index of each letter in
the current word, so that the current cube class (for styling)
is detected from the last element of this array, and similarly
the selected cubes class from the other elements of the array.

It remains to update the Game logic to detect when a cube has been
clicked twice in a row and deselect the cube (rather than throwing
an error due to the cube being re-used).