Implement de-select of Cube
This commit allows for the most recently clicked cube to be de-selected,
resulting in the styling and currentWord being correctly updated. To
accommodate this change, we remove the SelectedCubes context,
instead passing directly the CurrentWord to each Cube and
pulling out the (ordered) array of letterData.index.

For simplicity, we re-include the conditional of
`currentWord.length === 0` so that the later check of the index
of the last element of currentWord can ignore the case
where the length is 0 (and similarly for checking distance).