Improve Solution efficiency
This commit improves the time required to find all valid boggle
words; initial estimate is a reduction in time to approx 20%.

This is achieved through two changes.

First, the dictionary only has words up to size 12, so there
is no need to check the board for words of size 13 or
higher. Strictly speaking, this change does not contribute to
the improvement, but rather allows for the un-improved time
to not cause an out-of-memory error.

Second, the dictionary is now sorted; it is an Object with keys
integers i and values an Object of words of length i. This
improves the efficiency by allowing a reasonable implementation
of a search, by which both valid words of length i are tracked,
as well as letter combinations of length i
which could be the start of a word of length i + 1 or larger.
The search for such words starts at length i + 1, and halts as
soon as a word is found.