Add Solution display
Generating a solution now adds the found words to WordList (with
no concern for whether the words were already found). This change
is meant to help with building the vetical scroll feature
that is the goal of #2.

To this end, overflow-y attributes have been added to App.css; these
appear to have no effect (or at least not the desired effect) at
this time.

For ease of locating parent containers, the div which
contains the Board and Solution components now has a className of
BoardContainer.