Add broken scroll to WordList
This commit adds a partially-working scroll to the
WordList component through three steps:
1. Add overflow-x and -y: inherit to the css reset
2. Add overflow-x and -y: hidden to body
3. Add a fixed height to the WordList component

The final step is, at this time, implemented for the
'@media screen and (orientation: portrait)' query, and calculated as
(height of app) - (height of board) - (height of ButtonContainer) -
  (height of GenerateSolution button)

The last two terms contribute 58px total.

This does indeed add a scroll, but there is some issue with
the WordList component; the top of the scroll is perhaps
only the middle of the word list, and (since giving the
word list a fixed height), the first word (manually) entered
appears in the middle.

Check Flexbox align/justify flags?