Correct mobile landscape/portrait
This commit addresses the responsive display issues for
mobile devices.

1. The Interface-portrait class
is removed in favor of the media query orientation.
2. The Game component now has its width capped at 800px
in landscape mode.
3. In landscape, the Board is no longer a sqaure; rather
we sacrifice 18 px vertically to allow the Solution button
to display.
4. The Interface now uses the remaining space of the
App in landscape (width App - width Board).

It remains to verify that these changes work for the app
in desktop browsers of various window sizes.